Create HospitalERsystem
StackAsMyLinkedList and 
QueueAsMyLinkedList implementations demonstrated through a real-life hospital 
emergency room example. The aim of this assignment is to create the HospitalERSystem 
to provide a practical demonstration of how custom data structures, specifically, a stack 
and a queue implemented using linked lists can be used to manage real-world 
operations in a hospital setting. 
 Queue: Manages patients in the order they arrive, handling normal and 
emergency cases fairly 
 Stack: Handles undo operations to reverse the most recent treatments if a 
mistake was made, simulating real-life decision corrections.